User.save() now throws ClassNotFoundException & SQLException
Severe Exceptions such as ClassNotFoundException & SQLException
should not let the application continue to work. The application
should crash and give a verbose error message to help debugging.
In production, the verbosity for such exceptions could be reduced,
but we still don't want users to continue using the application
if severe errors occur.