absolute path for loginRedirect() + remove User empty constructor
When the user was not logged in, he was redirected from the index
to "login.xhtml". However, when running the project from the
base URL, i.e., localhost:8080/tps/, it was not redirected to
localhost:8080/tps/faces/login.xhtml but to
localhost:8080/tps/login.xhtml. Thus, the request was not handled
by jsf, and the page could not be displayed correctly.
To avoid this problem, the redirection path is now absolute:
/tps/faces/login.xhtml.
It does not seem like the best possible solution, but it works for
now.